+++ wl_game.cpp
@@ -21,6 +21,7 @@
 // WL_GAME.C
 
 #include <math.h>
+#include <stdlib.h>
 #include "version.h"
 #include "wl_def.h"
 #include "wl_act1.h"
@@ -996,7 +997,7 @@ void LatchNumberHERE (int x, int y, unsigned width, int32_t number)
     unsigned length,c;
     char str[20];
 
-    ltoa (number,str,10);
+    ltoa ((long int)number,str,10);
 
     length = (unsigned) strlen (str);
